requirements.txt
fastapi==0.116.1
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.1.1
langchain==0.3.3
langchain-openai==0.2.2
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4